+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#21](https://github.com/LaxarJS/laxar-mocks/issues/21): respect flush/applyViewChanges interaction
+
 
 ## v1.0.0-alpha.0